feat: Implement calendar and school cycle management pages
- Added detailed calendar page with edit and delete functionality.
- Created a form for creating new calendars.
- Developed a main calendar overview page with a table of events.
- Implemented edit functionality for school cycles.
- Added a detailed view for school cycles with edit and delete options.
- Created a form for adding new school cycles.
- Developed a main page for managing school cycles with a table view.
- Updated sidebar navigation to reflect new routes for calendars and school cycles.
- Enhanced breadcrumb navigation for better user experience.
- Integrated loading states and error handling in various components.
