+++ apps/system/app/escuela/[slug]/calendarios/[id]/page.tsx
