+++ apps/system/app/escuela/[slug]/ciclosEscolares/create/page.tsx
@@ -1,13 +1,12 @@
 "use client";
 
-import { use, useEffect, useState } from "react";
-import { useMutation, useQuery } from "convex/react";
+import { useEffect, useState } from "react";
+import { useMutation } from "convex/react";
 import { api } from "@/convex/_generated/api";
-import { useRouter } from "next/navigation";
-import { Button } from "@repo/ui/components/shadcn/button";
-import { Input } from "@repo/ui/components/shadcn/input";
-import { Card, CardContent, CardFooter, CardHeader, CardTitle } from "@repo/ui/components/shadcn/card";
-import { ArrowLeft } from "lucide-react";
+import { useParams, useRouter } from "next/navigation";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Card, CardContent, CardFooter, CardHeader, CardTitle, } from "@repo/ui/components/shadcn/card";
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@repo/ui/components/shadcn/form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { toast } from "sonner";
@@ -17,15 +16,12 @@ import { CicloEscolarFormValues, cicloEscolarSchema } from "@/app/shemas/cicloEs
 import { useBreadcrumbStore } from "@/app/store/breadcrumbStore";
 import { useEscuela } from "@/app/store/useEscuela";
 
-export default function EditarCicloEscolarPage({ params }: { params: Promise<{ id: string }> }) {
-    const { id } = use(params);
-    const idCicloEscolar = id as Id<"ciclosEscolares">;
+export default function CrearCicloEscolarPage() {
     const router = useRouter();
-    const actualizarCicloEscolar = useMutation(api.ciclosEscolares.actualizarCicloEscolar);
-    const [isSubmitting, setIsSubmitting] = useState(false);
-    const setItems = useBreadcrumbStore(state => state.setItems);
     const escuela = useEscuela((s) => s.escuela);
-    const cicloEscolar = useQuery(api.ciclosEscolares.obtenerCicloEscolarPorId, { cicloId: idCicloEscolar, escuelaId: escuela?._id as Id<"escuelas">  });
+    const crearCicloEscolar = useMutation(api.ciclosEscolares.crearCicloEscolar);
+    const params = useParams();
+    const slug = typeof params?.slug === "string" ? params.slug : "";
 
     const form = useForm<CicloEscolarFormValues>({
         resolver: zodResolver(cicloEscolarSchema),
@@ -36,29 +32,31 @@ export default function EditarCicloEscolarPage({ params }: { params: Promise<{ i
         }
     });
 
+    const [isSubmitting, setIsSubmitting] = useState(false);
+    const setItems = useBreadcrumbStore(state => state.setItems)
+
     useEffect(() => {
-        if (cicloEscolar && escuela) {
+        if (escuela) {
             setItems([
-                { label: `${escuela?.nombre}`, href: '/' },
+                { label: `${escuela?.nombre}`, href: `/escuela/${slug}` },
                 { label: 'Ciclos Escolares', href: '/ciclosEscolares' },
-                { label: `${cicloEscolar?.nombre}`, href: `/ciclosEscolares/${cicloEscolar._id}` },
-                { label: 'Editar', isCurrentPage: true },
-            ]);
+                { label: 'Crear Ciclo Escolar', isCurrentPage: true },
+            ])
         }
-    }, [escuela, cicloEscolar, setItems]);
+    }, [escuela, setItems, slug])
 
     const onSubmit = async (values: CicloEscolarFormValues) => {
         try {
             setIsSubmitting(true);
-            await actualizarCicloEscolar({
-                cicloId: idCicloEscolar,
+            await crearCicloEscolar({
                 escuelaId: escuela?._id as Id<"escuelas">,
                 nombre: values.nombre,
-                fechaInicio: Number(values.fechaInicio),
-                fechaFin: Number(values.fechaFin)
+                fechaInicio: new Date(values.fechaInicio).getTime(),
+                fechaFin: new Date(values.fechaFin).getTime()
             });
-            toast.success("Ciclo Escolar actualizada", { description: "La ciclo escolar se ha actualizado correctamente" });
-            router.push("/ciclosEscolares");
+            toast.success("Ciclo Escolar creada", { description: "La ciclo escolar se ha creado correctamente" });
+            router.push(`/${slug}/ciclosEscolares`);
+
         } catch (error) {
             toast.error("Error", {
                 description: "Ocurrió un error al guardar la ciclo escolar"
@@ -73,22 +71,20 @@ export default function EditarCicloEscolarPage({ params }: { params: Promise<{ i
         <div className="container px-4 sm:px-6 lg:px-8 py-10 mx-auto">
             <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4 mb-8">
                 <div className="flex items-center gap-2">
-                    <Button variant="outline" size="icon" onClick={() => router.back()}>
-                        <ArrowLeft className="h-4 w-4" />
-                    </Button>
                     <h1 className="text-2xl sm:text-3xl font-bold">
-                        Editar Ciclo Escolar
+                        Crear Nueva Ciclo Escolar
                     </h1>
                 </div>
             </div>
 
             <Card className="w-full max-w-2xl mx-auto">
                 <CardHeader>
-                    <CardTitle className="font-semibold text-center">Información de la ciclo escolar</CardTitle>
+                    <CardTitle className="font-semibold text-center">Información la Ciclo Escolar</CardTitle>
                 </CardHeader>
                 <Form {...form}>
                     <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-4">
                         <CardContent className="grid grid-cols-1 gap-6">
+
                             <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
                                 <FormField
                                     control={form.control}
@@ -132,6 +128,7 @@ export default function EditarCicloEscolarPage({ params }: { params: Promise<{ i
                                     )}
                                 />
                             </div>
+
                         </CardContent>
 
                         <CardFooter className="flex flex-col sm:flex-row justify-between gap-4 mt-4">
@@ -149,12 +146,13 @@ export default function EditarCicloEscolarPage({ params }: { params: Promise<{ i
                                 disabled={isSubmitting}
                                 className="w-full sm:w-auto"
                             >
-                                {isSubmitting ? "Actualizando..." : "Guardar Cambios"}
+                                {isSubmitting ? "Creando..." : "Crear Ciclo Escolar"}
                             </Button>
                         </CardFooter>
                     </form>
                 </Form>
             </Card>
         </div>
+
     );
 }
