+++ apps/system/app/escuela/[slug]/ciclosEscolares/page.tsx
