+++ apps/system/app/escuela/[slug]/page.tsx
